deprecate process_mth5 in favour of process_mth5_new, and update related imports
